Add some operand types.
To compile MOV we need to know the operand types. Thinking of keying
the hash map using something like (mnemonic, OpCodeType). Other
alternative would be to do the operand type logic in a method pointed
to by the OpCode table entry. Would like to keep those methods as
simple as possible though.